chore: rename twoFactor → two-factor, consolidate docs/internal context files
- src/components/twoFactor/ → two-factor/ (kebab-case consistency)
- Update 3 import paths (LockScreen, SettingsPage, SecurityStep)
- Move loose context dumps in docs/internal/ into context-bundles/ subdir
- Remove scripts/.env (plain-text credential, gitignored, rotate the key)
